hide memory usage behind an ifdef: the code needs to be modified for windows
